Adds no review info message and separates the modules
