+++ app/frontend/src/Components/GameDetails/Review/ReviewInput.module.scss
@@ -1,0 +1,36 @@
+@import "../../../colors";
+
+.review-input-container {
+    display: flex;
+    flex-direction: column;
+    width: 350px;
+    color: $color-text-light ;
+    background-color: $blue-green;
+    border-radius: 0.5em;
+    margin: 5px 5px 5px 5px;
+    overflow: hidden;
+    padding: 5px;
+    gap: 5px;
+    justify-content: center;
+
+    .header {
+        display: flex;
+        flex-direction: column;
+        gap: 7px;
+        margin: 5px 0px 0px 10px;
+        font-size: 18px;
+    }
+
+    .content-input {
+        background-color: $prussian-blue-dark-10 !important;
+        flex: 4;
+        width: 100%;
+        border-radius: 0.5em;
+        padding: 5px;
+        font-size: 13px;
+    }
+
+    .button {
+        color: $celadon;
+    }
+}
